+++ src/app/router.js
@@ -1,8 +1,8 @@
 import { Router } from 'express';
 import dotenv from 'dotenv';
-import CombinationController from './app/controllers/CombinationController.js'
-import { ControllerCombinationsById } from '../src/app/controllers/controller-combinations-by-id.js'
-import { ControllerTopCombinations } from '../src/app/controllers/controller-top-combinations.js'
+import CombinationController from '../presentation/controllers/CombinationController.js'
+import { ControllerCombinationsById } from '../presentation/controllers/controller-combinations-by-id.js'
+import { ControllerTopCombinations } from '../presentation/controllers/controller-top-combinations.js'
  dotenv.config();
 
 const Routes = new Router();
@@ -14,3 +14,4 @@ Routes.get('/combinations-by-id', ControllerCombinationsById.handle)
 Routes.get('/store-top-combinations', ControllerTopCombinations.handle)
 
 export default Routes;
+
